Adds 'context' parameter to k8s_object
This commit introduces a new parameter to k8s_object, in order
to manually specify the 'context' from the local kubeconfig file.
This is necessary in addition to the 'cluster' parameter in cases
where the bazel rule should be applied into a 'context' which is
configured on the local machine, but not currently selected.